+++ src/lib/core/models/index.ts
@@ -6,3 +6,4 @@ export * from './button.models';
 export * from './select-list.models';
 export * from './data-table.models';
 export * from './stepper.models';
+export * from './timeline.models';